+++ src/GitDevInsights/CodeInsights/Service/GeneratorFileExtensionChartService.php
@@ -3,10 +3,9 @@
 
 namespace GitDevInsights\CodeInsights\Service;
 
-use GitDevInsights\CodeInsights\Output\FileExtensionChartHTMLFileGenerator;
-use GitDevInsights\CodeInsights\Output\LanguageChartHTMLFileGenerator;
 use GitDevInsights\CodeInsights\Persistence\ProjectConfigDataProvider;
 use GitDevInsights\CodeInsights\Results\AnalysisResult;
+use GitDevInsights\TrendGraph\BasicStatistics\FileExtensionChartHTMLFileGenerator;
 
 class GeneratorFileExtensionChartService
 {
