+++ src/GitDevInsights/TrendGraph/BasicStatistics/FileExtensionChartHTMLFileGenerator.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace GitDevInsights\CodeInsights\Output;
+namespace GitDevInsights\TrendGraph\BasicStatistics;
 
-class FocusChartHTMLFileGenerator
+class FileExtensionChartHTMLFileGenerator
 {
     private array $jsonData;
 
     private string $chartTitle;
 
-    private string $filterKey = 'language-focus-data';
+    private string $filterKey = 'language-fileext-data';
 
     public function __construct(array $jsonData, string $chartTitle)
     {
@@ -57,15 +57,14 @@ class FocusChartHTMLFileGenerator
 
     public function renderChartOutput(): string
     {
+        // Extract dates and labels
         $dates = array_keys(array_reverse($this->jsonData[$this->filterKey]));
         $labels = array_keys($this->jsonData[$this->filterKey][$dates[0]]);
 
         // Generate datasets for each label
         $datasets = [];
         foreach ($labels as $label) {
-
             $data = [];
-
             foreach ($dates as $date) {
                 $data[] = $this->jsonData[$this->filterKey][$date][$label];
             }
@@ -73,9 +72,9 @@ class FocusChartHTMLFileGenerator
             $datasets[] = [
                 'label' => $label,
                 'data' => $data,
-                'backgroundColor' => 'rgba(' . rand(0, 255) . ', ' . rand(0, 255) . ', ' . rand(0, 255) . ', 0.5)',
+                'fill' => false,
                 'borderColor' => 'rgb(' . rand(0, 255) . ', ' . rand(0, 255) . ', ' . rand(0, 255) . ')',
-                'borderWidth' => 1,
+                'tension' => 0.1,
             ];
         }
 
@@ -109,7 +108,7 @@ class FocusChartHTMLFileGenerator
 </head>
 <body>
 <h1>'.$this->chartTitle.' - Programming Language Trend Analysis</h1>
-<h2>Backend / Frontend usage bye programming languages (Lines of Code)</h2>
+<h2>Usage of File extensions by Number of Lines of Code</h2>
 <div style="width: 80vw; height: 90vh; margin: 0 auto;">
     <canvas id="trend_chart" width="" height=""></canvas>
 </div>
@@ -121,7 +120,7 @@ class FocusChartHTMLFileGenerator
     // Create the chart
     var ctx = document.getElementById("trend_chart").getContext("2d");
     var myChart = new Chart(ctx, {
-        type: "bar",
+        type: "line",
         data: {
             labels: labels,
             datasets: datasets,
@@ -129,14 +128,13 @@ class FocusChartHTMLFileGenerator
         options: {
             scales: {
                 x: {
-                    stacked: true, // Enable stacking for X-axis
+                    type: "category",
                     title: {
                         display: true,
                         text: "Date"
                     }
                 },
                 y: {
-                    stacked: true, // Enable stacking for Y-axis
                     beginAtZero: true,
                     title: {
                         display: true,
@@ -159,7 +157,7 @@ class FocusChartHTMLFileGenerator
         return $html;
     }
 
-    public function writeChartOutputToFile($fileName) : void
+    public function writeChartOutputToFile(string $fileName) : void
     {
         $html = $this->renderChartOutput();
         file_put_contents($fileName, $html);
