+++ src/GitDevInsights/TrendGraph/BasicStatistics/FocusChartHTMLFileGenerator.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace GitDevInsights\CodeInsights\Output;
+namespace GitDevInsights\TrendGraph\BasicStatistics;
 
-class FileExtensionChartHTMLFileGenerator
+class FocusChartHTMLFileGenerator
 {
     private array $jsonData;
 
     private string $chartTitle;
 
-    private string $filterKey = 'language-fileext-data';
+    private string $filterKey = 'language-focus-data';
 
     public function __construct(array $jsonData, string $chartTitle)
     {
@@ -57,14 +57,15 @@ class FileExtensionChartHTMLFileGenerator
 
     public function renderChartOutput(): string
     {
-        // Extract dates and labels
         $dates = array_keys(array_reverse($this->jsonData[$this->filterKey]));
         $labels = array_keys($this->jsonData[$this->filterKey][$dates[0]]);
 
         // Generate datasets for each label
         $datasets = [];
         foreach ($labels as $label) {
+
             $data = [];
+
             foreach ($dates as $date) {
                 $data[] = $this->jsonData[$this->filterKey][$date][$label];
             }
@@ -72,9 +73,9 @@ class FileExtensionChartHTMLFileGenerator
             $datasets[] = [
                 'label' => $label,
                 'data' => $data,
-                'fill' => false,
+                'backgroundColor' => 'rgba(' . rand(0, 255) . ', ' . rand(0, 255) . ', ' . rand(0, 255) . ', 0.5)',
                 'borderColor' => 'rgb(' . rand(0, 255) . ', ' . rand(0, 255) . ', ' . rand(0, 255) . ')',
-                'tension' => 0.1,
+                'borderWidth' => 1,
             ];
         }
 
@@ -108,7 +109,7 @@ class FileExtensionChartHTMLFileGenerator
 </head>
 <body>
 <h1>'.$this->chartTitle.' - Programming Language Trend Analysis</h1>
-<h2>Usage of File extensions by Number of Lines of Code</h2>
+<h2>Backend / Frontend usage bye programming languages (Lines of Code)</h2>
 <div style="width: 80vw; height: 90vh; margin: 0 auto;">
     <canvas id="trend_chart" width="" height=""></canvas>
 </div>
@@ -120,7 +121,7 @@ class FileExtensionChartHTMLFileGenerator
     // Create the chart
     var ctx = document.getElementById("trend_chart").getContext("2d");
     var myChart = new Chart(ctx, {
-        type: "line",
+        type: "bar",
         data: {
             labels: labels,
             datasets: datasets,
@@ -128,13 +129,14 @@ class FileExtensionChartHTMLFileGenerator
         options: {
             scales: {
                 x: {
-                    type: "category",
+                    stacked: true, // Enable stacking for X-axis
                     title: {
                         display: true,
                         text: "Date"
                     }
                 },
                 y: {
+                    stacked: true, // Enable stacking for Y-axis
                     beginAtZero: true,
                     title: {
                         display: true,
@@ -157,7 +159,7 @@ class FileExtensionChartHTMLFileGenerator
         return $html;
     }
 
-    public function writeChartOutputToFile(string $fileName) : void
+    public function writeChartOutputToFile($fileName) : void
     {
         $html = $this->renderChartOutput();
         file_put_contents($fileName, $html);
